Be more careful with github when syncing.
This will log the error, but return gracefully, if we cannot find an
organization or user that we had seen before.

If someone had an organization, but then renamed it to something else.
We would have cached the organization with the old name, but then tried
to go query github for any new repos on it (using the old name).

This is a half-measure that protects against 404 errors in that case.

/cc @LuisAlejandro